Allow custom MailSettings to be applied
Handles the mail_settings attribute in SendgridBackend._build_sg_mail.

Note: if sandbox_mode is set in the supplied settings, its value is
overwritten with the default defined in this library's documentation to
preserve as much existing behavior as possible. But this can be changed if
desired to allow the user to provide a per-message sandbox_mode value in the
supplied MailSettings.